layout/forms/test/test_bug536567.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=536567
-->
<head>
  <title>Test for Bug 536567</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=536567">Mozilla Bug 536567</a>
<p id="display"></p>
<iframe id="content"></iframe>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 536567 **/

const Cc = Components.classes;
const Ci = Components.interfaces;
const Cu = Components.utils;
const Cm = Components.manager;

var ioSvc = Cc["@mozilla.org/network/io-service;1"].getService(Ci.nsIIOService);
var dirSvc = Cc["@mozilla.org/file/directory_service;1"].getService(Ci.nsIProperties);
var prefSvc = Cc["@mozilla.org/preferences-service;1"].getService(Ci.nsIPrefBranch);
var obSvc = Cc["@mozilla.org/observer-service;1"].getService(Ci.nsIObserverService);
if ("@mozilla.org/privatebrowsing;1" in Cc) {
  var pbSvc = Cc["@mozilla.org/privatebrowsing;1"].getService(Ci.nsIPrivateBrowsingService);
}

var tmpDir = dirSvc.get("TmpD", Ci.nsILocalFile);
var homeDir = dirSvc.get("Desk", Ci.nsILocalFile);
prefSvc.setBoolPref("browser.privatebrowsing.keep_current_session", true);

function newDir() {
  var dir = tmpDir.clone();
  dir.append("testdir" + Math.floor(Math.random() * 10000));
  dir.QueryInterface(Ci.nsILocalFile);
  dir.createUnique(Ci.nsIFile.DIRECTORY_TYPE, 0700);
  return dir;
}

var dirs = [];
for(var i = 0; i < 6; i++) {
  dirs.push(newDir());
}
dirs.push(homeDir);
var domains = ['http://mochi.test:8888', 'http://example.org:80', 'http://example.com:80'];
/*
 * These tests take 3 args each:
 * - which domain to load
 * - the filePicker displayDirectory we expect to be set
 * - the file to pick (in most cases this will show up in the next test,
 *   as indicated by the comments)
 */
var tests = [
  [0, 6, 0],//-> 3
  [1, 6, 1],//-> 4
  [2, 6, 2],//-> 5
  [0, 0, 3], //-> 6
  [1, 1, 1], //-> 8
  [2, 2, 2], //-> 9
  [0, 3, 1], //-> 7
  [0, 1, 0], //-> x
  [1, 1, 1], //-> x
  [2, 2, 2], //-> x
  "clear history",
  [0, 6, 0],//->15
  [1, 6, 1],//->16
  [2, 6, 2],//->17
  "pb on",
  [0, 0, 3], //-> 18
  [1, 1, 4], //-> 19
  [2, 2, 5], //-> 20
  [0, 3, 3], //-> x
  [1, 4, 4], //-> x
  [2, 5, 5], //-> x
  "pb off",
  [0, 0, 5], //-> 26
  [1, 1, 4], //-> 27
  [2, 2, 3], //-> 28
  "pb on",
  [0, 5, 5], //-> x
  [1, 4, 4], //-> x
  [2, 3, 3], //-> x
  "clear history",
  [0, 6, 0],
  [1, 6, 1],
  [2, 6, 2],
  "pb off"
];

var index;
var testIndex = 0;
var content = document.getElementById('content');

function runTest() {
  var test = tests[testIndex];
  if (test == undefined) {
    endTest();
  } else if (test == "pb on") {
    if (!pbSvc) {
      // private browsing service not available, finish test
      endTest();
      return;
    }
    pbSvc.privateBrowsingEnabled = true;
    testIndex++;
    runTest();
  } else if (test == "pb off") {
    pbSvc.privateBrowsingEnabled = false;
    testIndex++;
    runTest();
  } else if (test == "clear history") {
    obSvc.notifyObservers(null, "browser:purge-session-history", "");
    testIndex++;
    runTest();
  } else {
    index = test[2];
    content.setAttribute ('src', domains[test[0]] + '/tests/layout/forms/test/bug536567_subframe.html');
  }
}

function endTest() {
    prefSvc.clearUserPref("browser.privatebrowsing.keep_current_session");
    for(var i = 0; i < dirs.length - 1; i++) {
      dirs[i].remove(true);
    }

    SimpleTest.finish();
}

window.addEventListener("message", function(event) {
  var test = tests[testIndex];
  var returned = -1;
  for (var i = 0; i < dirs.length; i++) {
    if (dirs[i].path == event.data) {
      returned = i;
      break;
    }
  }
  if (test[1] == -1) {
    ok(false, "We should never get an unknown directory back");
  } else {
    is(returned, test[1], 'test ' + testIndex);
  }
  
  testIndex++;
  runTest();
}, false);

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  setTimeout(runTest, 0);
};

</script>
</pre>
</body>
</html>
